remove pinconfig.
