+++ wroprg/src/hardware/measurements.py
@@ -4,6 +4,10 @@ import logging
 import os
 import json
 from typing import Any, Dict, Optional
+import threading
+from collections import deque
+from base.shutdown_handling import ShutdownInterface
+from hardware.hardware_interface import HardwareInterface, RobotState
 
 logger = logging.getLogger(__name__)
 
@@ -137,3 +141,80 @@ class MeasurementsLogger:
         if self._file is not None:
             self._file.close()
             self._file = None
+
+class MeasurementFileLog(ShutdownInterface):
+    """Class to read and store measurements from hardware sensors in a separate thread."""
+
+    def __init__(self, hardware_interface: HardwareInterface):
+        self.measurements: deque[Measurement] = deque(maxlen=5)  # Store last 5 measurements
+        self._reading_thread: threading.Thread | None = None
+        self._stop_event = threading.Event()
+        self._hardware_interface = hardware_interface
+        self._mlogger = MeasurementsLogger()
+        self._rpi = hardware_interface._rpi
+
+    def add_measurement(self, measurement: Measurement) -> None:
+        """Add a new measurement to the list."""
+        self.measurements.append(measurement)
+        self._mlogger.write_measurement(measurement)
+        logger.debug("Added measurement: %s", measurement)
+
+    def add_comment(self, comment: str) -> None:
+        """Add a comment to the measurements log."""
+        self._mlogger.write_comment(comment)
+
+    def get_latest_measurement(self) -> Measurement | None:
+        """Get the latest measurement."""
+        if self.measurements:
+            return self.measurements[-1]
+        return None
+
+    def _read_hardware_loop(self) -> None:
+        """Thread target: read hardware every 0.5 seconds."""
+        start_time = time.time()
+        while not self._stop_event.is_set():
+            if self._hardware_interface is not None:
+                metrics: Dict[str, Any] = {}
+                state: RobotState = self._hardware_interface.read_state()
+                steering_angle = self._hardware_interface.get_steering_angle()
+                roll, pitch, yaw = self._hardware_interface.get_orientation()
+                # Create a new measurement with the current timestamp
+                timestamp = time.time()
+                counter = int((timestamp - start_time)*1000)
+
+                (_,_,_, metrics) = self._hardware_interface.camera_measurements.get_distance()
+
+                measurement = Measurement(state.left, state.right, state.front,
+                            steering_angle,
+                            roll, pitch, yaw,counter,extra_metrics=metrics)
+
+                self.add_measurement(measurement)
+                self._rpi.log_message(front=state.front, left=state.left,
+                                                           right=state.right,current_yaw=yaw,
+                                               current_steering=steering_angle)
+
+
+            time.sleep(0.25)
+
+    def start_reading(self) -> None:
+        """Start the background thread for reading hardware."""
+        self._mlogger.writeheader()  # Write header to the measurements file
+        if self._reading_thread is None or not self._reading_thread.is_alive():
+            self._stop_event.clear()
+            self._reading_thread = threading.Thread(target=self._read_hardware_loop, daemon=True)
+            self._reading_thread.start()
+            self._hardware_interface.disable_logger()
+
+    def stop_reading(self) -> None:
+        """Stop the background thread for reading hardware."""
+        self._stop_event.set()
+        if self._reading_thread is not None:
+            self._reading_thread.join()
+            self._reading_thread = None
+
+    def shutdown(self) -> None:
+        """Shutdown the reader and stop the reading thread."""
+        self.stop_reading()
+        self.measurements.clear()
+        self._mlogger.close_file()
+        logger.info("MeasurementsReader shutdown complete.")
